% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence}
\usage{
check_convergence(raw_draws)
}
\arguments{
\item{raw_draws}{(\code{dynIRT_draws}) A posterior draws}
}
\description{
Check convergence
}
