% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{check_item_labels}
\alias{check_item_labels}
\title{internal function to check \code{item_labels} argument}
\usage{
check_item_labels(outcomes_labeled_item, item_labels)
}
\description{
internal function to check \code{item_labels} argument
}
