% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-loo.r
\name{check_log_lik_finite}
\alias{check_log_lik_finite}
\title{Check log-likelihood draws for non-finite values}
\usage{
check_log_lik_finite(a, drop = FALSE, call = rlang::caller_env())
}
\arguments{
\item{a}{(array) Iteration by chain by group array of log-likelihoods.}

\item{drop}{(logical) Discard offending iterations rather than aborting?}

\item{call}{(environment) Frame to report errors against.}
}
\value{
\code{a}, possibly with iterations removed.
}
\description{
Check log-likelihood draws for non-finite values
}
\keyword{internal}
