% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.r
\name{copy_mixfac_attrs}
\alias{copy_mixfac_attrs}
\title{Copy mixfac label attributes from one object to another}
\usage{
copy_mixfac_attrs(to, from)
}
\description{
Copy mixfac label attributes from one object to another
}
\keyword{internal}
