% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{.mixfac_drop_default}
\alias{.mixfac_drop_default}
\title{Extract draws from fitted model}
\usage{
.mixfac_drop_default
}
\arguments{
\item{x}{(\code{dbmm_fitted} object) A fitted model produced by \code{fit()}.}

\item{drop}{(character vector) A regular expression (or logical scalar).
Parameters that match any of the regular expressions will be dropped.}

\item{format}{(string) The format of the returned draws or point
estimates. Must be a valid format from the ‘posterior’ package. The
default is \code{"df"}, which is what other \code{dbmm} functions will
expect, but other options include \code{"array"}, \code{"matrix"}, and \code{"list"}.}

\item{check}{(logical)}
}
\value{
Draws from the posterior distribution of the selected parameters.
}
\description{
Extract draws from fitted model
}
