% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{extract}
\alias{extract}
\title{Extract object from a stan output}
\usage{
extract(fit, drop_rex = "^z_", format = "df")
}
\arguments{
\item{format}{}
}
\description{
Extract object from a stan output
}
