% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract object from a stan output}
\usage{
extract_draws(fit, drop_rex = "^z_", format = "df")
}
\arguments{
\item{format}{}
}
\value{
Draws
}
\description{
Extract object from a stan output
}
