% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynirt.R
\name{fit}
\alias{fit}
\title{dynIRT main function}
\usage{
fit(
  data,
  chains = 4,
  parallelize_within_chains = FALSE,
  threads_per_chain = NULL,
  force_recompile = FALSE,
  init_kappa = FALSE,
  init = NULL,
  return_data = TRUE,
  n_dim = 1,
  constant_alpha = FALSE,
  separate_eta = TRUE,
  lambda_zeros = NULL,
  df_sigma_metric = 4,
  df_sigma_alpha_evol = 4,
  df_sigma_eta_evol = 4,
  mu_sigma_metric = 0.5,
  mu_sigma_alpha_evol = 0.1,
  mu_sigma_eta_evol = 0.1,
  sd_sigma_metric = 0.5,
  sd_sigma_alpha_evol = 0.1,
  sd_sigma_eta_evol = 0.1,
  seed = 123,
  ...
)
}
\arguments{
\item{seed}{An integer vector of length one indicating the state of Stan’s pseudo-random number generator}

\item{...}{}
}
\value{
A dynIRT object containing
\describe{
\item{unit_labels}
\item{time_labels}
\item{binary_item_labels}
\item{ordinal_item_labels}
\item{metric_item_labels}
}
}
\description{
Fit a dynIRT model using stan
}
