% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify}
\alias{identify}
\title{Identify the sign and rotation of the output}
\usage{
identify(
  raw_draws,
  rotate = FALSE,
  varimax = TRUE,
  normalize = TRUE,
  id_with = NULL
)
}
\arguments{
\item{id_with}{}
}
\description{
Identify the sign and rotation of the output
}
