% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_draws}
\alias{identify_draws}
\title{Identify the sign and rotation of the parameter draws}
\usage{
identify_draws(
  raw_draws,
  rotate = FALSE,
  varimax = TRUE,
  normalize = TRUE,
  id_with = NULL,
  sign = NULL,
  check = TRUE
)
}
\arguments{
\item{raw_draws}{(\code{draws_df}) A posterior draws}

\item{check}{(logical)}
}
\value{
Identified draws
}
\description{
Identify the sign and rotation of the parameter draws
}
