% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_rotation}
\alias{identify_rotation}
\title{Identify the rotation of the output}
\usage{
identify_rotation(raw_draws, varimax, targets, normalize, item_type)
}
\arguments{
\item{raw_draws}{(\code{draws_df}) Posterior draws}

\item{varimax}{(logical) Should a varimax rotation be applied within each
draw? Defaults to \code{TRUE}.}

\item{normalize}{(logical) Should Kaiser normalization be performed before
varimax rotation?}

\item{item_type}{(string) The item type. Default is "binary".}
}
\value{
Rotated object
}
\description{
Identify the rotation of the output
}
