% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_sign}
\alias{identify_sign}
\title{Identify signs}
\usage{
identify_sign(raw_draws, sign)
}
\arguments{
\item{raw_draws}{(\code{draws_df}) A posterior draws}

\item{sign}{(integer) Should the sign of the average identified loading be
negative (\code{-1}) or positive (\code{+1}).}
}
\value{
A \code{dynIRT_labeled} object.
}
\description{
Identify signs
}
