% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{label_draws}
\alias{label_draws}
\title{Label the output}
\usage{
label_draws(draws, regex_pars = NULL, check = TRUE)
}
\arguments{
\item{draws}{(\code{dbmm_identified}) An identified dbmm fitted object}

\item{regex_pars}{(named character vector) Vector of regular expressions, to
be used to pick out columns of \code{draws}. The names of this vector will be
used to name the resulting list of data frames. If \code{NULL} (the default),
all parameters will be returned.}

\item{check}{(logical) Should the class of \code{draws} be checked? Defaults to
\code{TRUE}.}
}
\value{
A list
}
\description{
Label the output
}
