% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{label_mixfac}
\alias{label_mixfac}
\title{Exported function}
\usage{
label_mixfac(x, make_long = FALSE)
}
\arguments{
\item{x}{An \code{draws_rvars} object, such as the output of \code{identify_mixfac()}.}

\item{make_long}{Should the output be a long data frame? (Requires dropping variables)}
}
\description{
Exported function
}
