% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{label_modgirt}
\alias{label_modgirt}
\title{Attach unit, period, and item labels to modgirt draws}
\usage{
label_modgirt(x, check = TRUE)
}
\arguments{
\item{x}{A \code{draws_rvars} object, typically from \code{\link[=identify_modgirt]{identify_modgirt()}}.}

\item{check}{(logical) Check that \code{x} is a \code{draws_rvars} object?}
}
\value{
A \code{modgirt_lab} object.
}
\description{
Attach unit, period, and item labels to modgirt draws
}
