% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept}
\alias{plot_intercept}
\title{plot item intercepts}
\usage{
plot_intercept(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  maintitle = NULL
)
}
\arguments{
\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}
}
\value{
A plot showing the estimated intercept
}
\description{
plot item intercepts
}
