% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept2}
\alias{plot_intercept2}
\title{test function}
\usage{
plot_intercept2(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  maintitle = NULL,
  recode = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dynIRT_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}

\item{recode}{(string)}
}
\value{
A plot showing the estimated intercept
}
\description{
test function
}
