% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept3}
\alias{plot_intercept3}
\title{test function}
\usage{
plot_intercept3(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  item_type = c("metric", "binary", "ordinal"),
  maintitle = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dynIRT_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{item_type}{(string) Should "binary", "ordinal", or "metric" loadings be
used to visualize? Default is "metric".}

\item{maintitle}{(string) Title of the plot}
}
\value{
A plot showing the estimated intercept
}
\description{
test function
}
