% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_ave}
\alias{plot_scores_ave}
\title{plot average factor scores}
\usage{
plot_scores_ave(
  outcomes_labeled,
  xdim = "Dimension 1 (Valence)",
  ydim = "Dimension 2 (Spatial)"
)
}
\arguments{
\item{xdim}{(string) x-axis label for the plot}

\item{ydim}{(string) y-axis label for the plot}
}
\value{
plot
}
\description{
plot average factor scores
}
