% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_timetrend}
\alias{plot_scores_timetrend}
\title{plot time series factor scores}
\usage{
plot_scores_timetrend(
  outcomes_labeled,
  xdim = "Dimension 1 (Valence)",
  ydim = "Dimension 2 (Spatial)"
)
}
\arguments{
\item{xdim}{(string) x-axis label for the plot}

\item{ydim}{(string) y-axis label for the plot}
}
\value{
plot
}
\description{
plot time series factor scores
}
