% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{set_signs}
\alias{set_signs}
\title{Set Signs}
\usage{
set_signs(modgirt_rvar, signs = 1)
}
\arguments{
\item{modgirt_rvar}{The model object containing the parameters.}

\item{signs}{A vector of signs to be applied to the parameters.
Scalar values are allowed and will be recycled. Default is 1.}
}
\value{
A modified model object with the signs of the parameters updated.
}
\description{
This function sets the signs of the parameters in a model based on
user-defined signs.
}
\details{
This function sets the signs of the parameters in the model object
\code{modgirt_rvar} based on the user-defined signs provided in the
\code{signs} argument. The function applies the sign flips to the parameters
and returns a modified model object with the updated signs.
}
\examples{
# Set signs to -1 for all parameters
set_signs(modgirt_rvar, signs = -1)

# Set signs to 1 for all parameters
set_signs(modgirt_rvar)

}
