% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{shape}
\alias{shape}
\title{prepare data for dynIRT}
\usage{
shape(
  long_data,
  unit_var,
  time_var,
  item_var,
  value_var,
  ordinal_items = NA,
  binary_items = NA,
  max_cats = 10,
  standardize = TRUE,
  make_indicator_for_zeros = TRUE,
  periods_to_estimate = NULL
)
}
\arguments{
\item{periods_to_estimate}{}
}
\value{
A list that is formatted for stan
}
\description{
prepare data for dynIRT
}
