% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{shape_data_modgirt}
\alias{shape_data_modgirt}
\title{shape_data_modgirt function}
\usage{
shape_data_modgirt(
  long_data,
  time_var,
  item_var,
  value_var,
  weight_var = NULL,
  periods_to_estimate
)
}
\arguments{
\item{long_data}{The long-format data frame containing the raw data.}

\item{time_var}{The name of the variable representing the time identifier.}

\item{item_var}{The name of the variable representing the item identifier.}

\item{value_var}{The name of the variable representing the repsonse value.}

\item{weight_var}{The name of the variable representing subject-specific
weights (optional).}

\item{periods_to_estimate}{The periods to estimate (optional).}

\item{unit_var}{The name of the variable representing the unit identifier.}
}
\value{
A list containing the dimensions of the data (T, G, Q, K, D), the
four-dimensional cross-tabulation (SSSS), the matrix of nonzero loadings
(beta_nonzero), and the matrix of signed loadings (beta_sign).
}
\description{
This function prepares the data for the MODGIRT analysis by creating a
four-dimensional cross-tabulation and setting up the necessary variables
and parameters.
}
