% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{sign_permute}
\alias{sign_permute}
\title{Permute signs}
\usage{
sign_permute(
  lambda_item,
  c_cur,
  lcols,
  item_type = c("binary", "ordinal", "metric")
)
}
\arguments{
\item{lambda_item}{lambda object}

\item{c_cur}{current row indicator}

\item{lcols}{column}

\item{item_type}{(string) The item type. Default is "binary".}
}
\value{
A sign-permuted object
}
\description{
Permute signs
}
