% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social_outcomes_2021}
\alias{social_outcomes_2021}
\title{Public opinion data from 2021}
\format{
A tibble with following objects:
\describe{
\item{st}{A character vector indicating state names}
\item{year}{A integer vector indicating year}
\item{outcome}{A character vector describing the issue area}
\item{value}{A numeric indicating the sentiment toward the corresponding issue}
}
}
\source{
Caughey and Warshaw
}
\usage{
social_outcomes_2021
}
\description{
Public opinion data from 2021
}
\keyword{datasets}
