% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{sort_mixfac}
\alias{sort_mixfac}
\title{order factors in a model based on sums of squared loadings}
\usage{
sort_mixfac(x, check = TRUE)
}
\arguments{
\item{mixfac_rvar}{A \code{draws_rvar} object from a mixed-factor model}
}
\value{
A \code{draws_rvar} object with factors ordered by explanatory power
}
\description{
This function takes a model based on posterior draws and orders the factors
based on their estimated sums of squares. Factors with larger sums of squares
will be placed first in the sort model.
}
