% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-utils.r
\name{with_seed}
\alias{with_seed}
\title{Evaluate code with a temporary RNG seed}
\usage{
with_seed(seed, code)
}
\arguments{
\item{seed}{(integer or \code{NULL}) Seed to set. If \code{NULL}, \code{code} is evaluated
with the ambient RNG state untouched.}

\item{code}{Code to evaluate.}
}
\description{
Evaluate code with a temporary RNG seed
}
\keyword{internal}
