% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{extract_draws}
\alias{extract_draws}
\title{Extract draws from fitted model}
\usage{
extract_draws(fit, drop_rex = "^z_", format = "df")
}
\arguments{
\item{fit}{(dynIRT object) A fitted model produced by \code{fit()}.}

\item{drop_rex}{(character vector) A vector of regular expressions.
Parameters that match any of the regular expressions will be dropped.}

\item{format}{(string) The format of the returned draws or point
estimates. Must be a valid format from the ‘posterior’ package. The
default is \code{"df"}, which is what other \code{dynIRTtest} functions will
expect, but other options include \code{"array"}, \code{"matrix"}, and \code{"list"}.}
}
\value{
Draws from the posterior distribution of the selected parameters.
}
\description{
Extract draws from fitted model
}
