% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{extract_mixfac_draws}
\alias{extract_mixfac_draws}
\title{Extract draws from fitted model}
\usage{
extract_mixfac_draws(x, drop = .mixfac_drop_default, check = TRUE)
}
\arguments{
\item{x}{(\code{dbmm_fitted} object) A fitted model produced by \code{fit()}.}

\item{drop}{(character vector) A regular expression (or logical scalar).
Parameters that match any of the regular expressions will be dropped.}

\item{check}{(logical)}
}
\value{
Draws from the posterior distribution of the selected parameters.
}
\description{
Extract draws from fitted model
}
