% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{extract_mixfac_draws}
\alias{extract_mixfac_draws}
\title{Extract draws from fitted model}
\usage{
extract_mixfac_draws(x, drop = "^z_|^chol", check = TRUE)
}
\arguments{
\item{x}{(\code{dbmm_fitted} object) A fitted model produced by \code{fit()}.}

\item{drop}{(character vector) A regular expression (or logical scalar).
Parameters that match any of the regular expressions will be dropped.}

\item{check}{(logical)}

\item{format}{(string) The format of the returned draws or point
estimates. Must be a valid format from the ‘posterior’ package. The
default is \code{"df"}, which is what other \code{dbmm} functions will
expect, but other options include \code{"array"}, \code{"matrix"}, and \code{"list"}.}
}
\value{
Draws from the posterior distribution of the selected parameters.
}
\description{
Extract draws from fitted model
}
