% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_modgirt}
\alias{fit_modgirt}
\title{Fit a multidimensional ordinal group-level IRT model using Stan}
\usage{
fit_modgirt(
  stan_data,
  chains = 4,
  return_data = TRUE,
  n_factor = 1,
  force_recompile = FALSE,
  signed_loadings,
  nonzero_loadings,
  link = "probit",
  seed = NULL,
  ...
)
}
\arguments{
\item{stan_data}{The data for the model.}

\item{chains}{The number of chains to run in parallel (default is 4).}

\item{return_data}{A logical value indicating whether to return the input
data along with the model fit (default is TRUE).}

\item{n_factor}{The number of factors in the model.}

\item{force_recompile}{A logical value indicating whether to force
recompilation of the Stan model (default is FALSE).}

\item{signed_loadings}{The signed loadings matrix.}

\item{nonzero_loadings}{The nonzero loadings matrix.}

\item{link}{The link function to use (default is "probit").}

\item{seed}{The random seed for reproducibility.}

\item{...}{Additional arguments to be passed to the Stan sampling function.}
}
\value{
A list containing the model fit and optionally the input data.
}
\description{
This function fits a MODGIRT model using Stan.
It takes in the following arguments:
}
