% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{identify_mixfac}
\alias{identify_mixfac}
\title{Exported function}
\usage{
identify_mixfac(
  x,
  method = "varimax",
  whiten = FALSE,
  ref_t = "last",
  identify_with_type
)
}
\description{
Exported function
}
