% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_modgirt}
\alias{identify_modgirt}
\title{Identify MODGIRT draws}
\usage{
identify_modgirt(x, method = "varimax")
}
\arguments{
\item{x}{A fitted MODGIRT model object or \code{draws_rvars} object}
}
\value{
A list containing the identified MODGIRT model parameters.
}
\description{
This function identifies the MODGIRT model by postprocessing the draws from
the posterior distribution.
}
