% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_modgirt}
\alias{identify_modgirt}
\title{Identify MODGIRT model}
\usage{
identify_modgirt(modgirt_fit)
}
\arguments{
\item{modgirt_fit}{The fitted MODGIRT model object.}
}
\value{
A list containing the identified MODGIRT model parameters.
}
\description{
This function identifies the MODGIRT model by postprocessing the draws from
the posterior distribution.
}
