% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{identify_rotation}
\alias{identify_rotation}
\title{Identify the rotation of the output}
\usage{
identify_rotation(raw_draws, varimax, normalize, id_with)
}
\arguments{
\item{id_with}{}
}
\value{
Rotated object
}
\description{
Identify the rotation of the output
}
