% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{label_draws}
\alias{label_draws}
\title{Label the output}
\usage{
label_draws(draws, regex_pars = NULL, check = TRUE)
}
\arguments{
\item{draws}{(\code{dynIRT_identified}) A identified dynIRT fitted object}

\item{check}{(logical)}
}
\value{
A list
}
\description{
Label the output
}
