% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-loo.r
\name{log_lik_group_key}
\alias{log_lik_group_key}
\title{Group log-likelihood columns}
\usage{
log_lik_group_key(idx, group)
}
\arguments{
\item{idx}{(data frame) Output of \code{\link[=log_lik_index]{log_lik_index()}}.}

\item{group}{(string) Grouping level.}
}
\value{
A factor, in order of first appearance, or \code{NULL} for
\code{"observation"}.
}
\description{
Group log-likelihood columns
}
\keyword{internal}
