% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept}
\alias{plot_intercept}
\title{plot item intercepts}
\usage{
plot_intercept(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  maintitle = NULL,
  item_labels = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dbmm_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}

\item{item_labels}{(string) Named string vector where each element represents the new labels and the names of the elements correspond to the original label in \code{outcomes_labeled}. Default is NULL.}
}
\value{
A plot showing the estimated intercept
}
\description{
plot item intercepts
}
