% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept2}
\alias{plot_intercept2}
\title{test function}
\usage{
plot_intercept2(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  maintitle = NULL
)
}
\description{
test function
}
