% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intercept3}
\alias{plot_intercept3}
\title{test function}
\usage{
plot_intercept3(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = "Estimated Intercept",
  id_with = c("metric", "binary", "ordinal"),
  maintitle = NULL
)
}
\description{
test function
}
