% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_loadings}
\alias{plot_loadings}
\title{plot item loadings}
\usage{
plot_loadings(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = NULL,
  maintitle = "Item Loadings"
)
}
\arguments{
\item{outcomes_labeled}{A \code{dynIRT_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot. Default is ``Item Loadings''}
}
\value{
A plot showing item loadings
}
\description{
plot item loadings
}
