% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_ave}
\alias{plot_scores_ave}
\title{plot average factor scores across different units}
\usage{
plot_scores_ave(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = NULL,
  maintitle = "Average Factor Scores",
  unit_labels = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dbmm_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot. Default is "Agerage Factor Scores".}

\item{unit_labels}{(string) Named string vector where each element represents the new unit labels and the names of the elements correspond to the original unit labels in \code{outcomes_labeled}. Default is NULL.}
}
\value{
A plot showing average factor scores
}
\description{
plot average factor scores across different units
}
