% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_ave}
\alias{plot_scores_ave}
\title{plot average factor scores}
\usage{
plot_scores_ave(outcomes_labeled)
}
\arguments{
\item{outcomes_labeled}{}
}
\value{
plot
}
\description{
plot average factor scores
}
