% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_ave}
\alias{plot_scores_ave}
\title{plot average factor scores}
\usage{
plot_scores_ave(
  outcomes_labeled,
  xtitle = NULL,
  ytitle = NULL,
  maintitle = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dynIRT_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}
}
\value{
A plot showing average factor scores
}
\description{
plot average factor scores
}
