% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_timetrend}
\alias{plot_scores_timetrend}
\title{plot time series factor scores}
\usage{
plot_scores_timetrend(outcomes_labeled)
}
\arguments{
\item{outcomes_labeled}{}
}
\value{
plot
}
\description{
plot time series factor scores
}
