% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_timetrend}
\alias{plot_scores_timetrend}
\title{Plot time series factor scores. Currently only support 2-dimensional plot.}
\usage{
plot_scores_timetrend(
  outcomes_labeled,
  xtitle = "Dimension 1",
  ytitle = "Dimension 2",
  maintitle = NULL,
  unit_labels = NULL
)
}
\arguments{
\item{outcomes_labeled}{A \code{dynIRT_labeled} object}

\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}

\item{unit_labels}{(string) Named string vector where each element represents the new unit labels and the names of the elements correspond to the original unit labels in \code{outcomes_labeled}. Default is NULL.}
}
\value{
A plot plot showing time trend
}
\description{
Plot time series factor scores. Currently only support 2-dimensional plot.
}
