% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_scores_timetrend}
\alias{plot_scores_timetrend}
\title{plot time series factor scores}
\usage{
plot_scores_timetrend(
  outcomes_labeled,
  xdim,
  ydim,
  xtitle = "Dimension 1",
  ytitle = "Dimension 2",
  maintitle = NULL
)
}
\arguments{
\item{xtitle}{(string) x-axis label of the plot}

\item{ytitle}{(string) y-axis label of the plot}

\item{maintitle}{(string) Title of the plot}
}
\value{
A plot plot showing time trend
}
\description{
plot time series factor scores
}
