% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_analysis.R
\name{rotate_modgirt}
\alias{rotate_modgirt}
\title{Apply rotation to MODGIRT draws}
\usage{
rotate_modgirt(modgirt_rvar, rotmat)
}
\arguments{
\item{modgirt_rvar}{A \code{draws_rvar} object from a MODGIRT model}

\item{rotat}{An I-by-D rotation matrix}
}
\value{
A \code{draws_rvar} object of rotated draws
}
\description{
This function applies the given rotation to each draw from the posterior
distribution of the MODGIRT parameters
}
\examples{
rotmat <- varimax(E(modgirt_signed$beta))$rotmat
modgirt_rotated <- rotate_modgirt(modgirt_signed, rotmat)

}
