% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-process_draws.r
\name{sign_mixfac}
\alias{sign_mixfac}
\title{Set Signs}
\usage{
sign_mixfac(x, signs = 1)
}
\arguments{
\item{x}{The model object containing the parameters.}

\item{signs}{A vector of signs to be applied to the parameters.
Scalar values are allowed and will be recycled. Default is 1.}
}
\value{
A modified model object with the signs of the parameters updated.
}
\description{
This function sets the signs of the parameters of a MIXFAC model based on
user-defined signs.
}
\details{
This function sets the signs of the parameters in the model object
\code{x} based on the user-defined signs provided in the
\code{signs} argument. The function applies the sign flips to the parameters
and returns a modified model object with the updated signs.
}
